+++ getData.js
@@ -1,515 +1,514 @@
-'use strict';
-
-export const getData = function () {
-    return [{
-            "idDrink": "17222",
-            "strDrink": "A1",
-            "strCategory": "Cocktail",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "Pour all ingredients into a cocktail shaker, mix and serve over ice into a chilled glass.",
-            "strInstructionsDE": "Alle Zutaten in einen Cocktailshaker geben, mischen und über Eis in ein gekühltes Glas servieren.",
-            "strInstructionsIT": "Versare tutti gli ingredienti in uno shaker, mescolare e servire con ghiaccio in un bicchiere freddo.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/2x8thr1504816928.jpg",
-            "strIngredient1": "Gin",
-            "strIngredient2": "Grand Marnier",
-            "strIngredient3": "Lemon Juice",
-            "strIngredient4": "Grenadine",
-            "strMeasure1": "1 3/4 shot ",
-            "strMeasure2": "1 Shot ",
-            "strMeasure3": "1/4 Shot",
-            "strMeasure4": "1/8 Shot",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-07 21:42:09"
+export const getData = function() {
+    return [
+        {
+          "idDrink": "17222",
+          "strDrink": "A1",
+          "strCategory": "Cocktail",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "Pour all ingredients into a cocktail shaker, mix and serve over ice into a chilled glass.",
+          "strInstructionsDE": "Alle Zutaten in einen Cocktailshaker geben, mischen und über Eis in ein gekühltes Glas servieren.",
+          "strInstructionsIT": "Versare tutti gli ingredienti in uno shaker, mescolare e servire con ghiaccio in un bicchiere freddo.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/2x8thr1504816928.jpg",
+          "strIngredient1": "Gin",
+          "strIngredient2": "Grand Marnier",
+          "strIngredient3": "Lemon Juice",
+          "strIngredient4": "Grenadine",
+          "strMeasure1": "1 3/4 shot ",
+          "strMeasure2": "1 Shot ",
+          "strMeasure3": "1/4 Shot",
+          "strMeasure4": "1/8 Shot",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-07 21:42:09"
         },
         {
-            "idDrink": "13501",
-            "strDrink": "ABC",
-            "strCategory": "Shot",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Shot glass",
-            "strInstructions": "Layered in a shot glass.",
-            "strInstructionsDE": "Schichtaufbau in einem Schnapsglas.",
-            "strInstructionsIT": "Versa in ordine di lettera i vari ingredienti. 1/3 del bicchiere va riempito con l'Amaretto, 1/3 di Baileys e il restante di Cognac.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/tqpvqp1472668328.jpg",
-            "strIngredient1": "Amaretto",
-            "strIngredient2": "Baileys irish cream",
-            "strIngredient3": "Cognac",
-            "strMeasure1": "1/3 ",
-            "strMeasure2": "1/3 ",
-            "strMeasure3": "1/3 ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2016-08-31 19:32:08"
+          "idDrink": "13501",
+          "strDrink": "ABC",
+          "strCategory": "Shot",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Shot glass",
+          "strInstructions": "Layered in a shot glass.",
+          "strInstructionsDE": "Schichtaufbau in einem Schnapsglas.",
+          "strInstructionsIT": "Versa in ordine di lettera i vari ingredienti. 1/3 del bicchiere va riempito con l'Amaretto, 1/3 di Baileys e il restante di Cognac.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/tqpvqp1472668328.jpg",
+          "strIngredient1": "Amaretto",
+          "strIngredient2": "Baileys irish cream",
+          "strIngredient3": "Cognac",
+          "strMeasure1": "1/3 ",
+          "strMeasure2": "1/3 ",
+          "strMeasure3": "1/3 ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2016-08-31 19:32:08"
         },
         {
-            "idDrink": "17225",
-            "strDrink": "Ace",
-            "strCategory": "Cocktail",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Martini Glass",
-            "strInstructions": "Shake all the ingredients in a cocktail shaker and ice then strain in a cold glass.",
-            "strInstructionsDE": "Alle Zutaten in einem Cocktailshaker mit Eis schütteln und dann in einem kalten Glas abseihen.",
-            "strInstructionsIT": "Shakerare tutti gli ingredienti in uno shaker e ghiaccio, quindi filtrare in un bicchiere freddo.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/l3cd7f1504818306.jpg",
-            "strIngredient1": "Gin",
-            "strIngredient2": "Grenadine",
-            "strIngredient3": "Heavy cream",
-            "strIngredient4": "Milk",
-            "strIngredient5": "Egg White",
-            "strMeasure1": "2 shots ",
-            "strMeasure2": "1/2 shot ",
-            "strMeasure3": "1/2 shot ",
-            "strMeasure4": "1/2 shot",
-            "strMeasure5": "1/2 Fresh",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-07 22:05:06"
+          "idDrink": "17225",
+          "strDrink": "Ace",
+          "strCategory": "Cocktail",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Martini Glass",
+          "strInstructions": "Shake all the ingredients in a cocktail shaker and ice then strain in a cold glass.",
+          "strInstructionsDE": "Alle Zutaten in einem Cocktailshaker mit Eis schütteln und dann in einem kalten Glas abseihen.",
+          "strInstructionsIT": "Shakerare tutti gli ingredienti in uno shaker e ghiaccio, quindi filtrare in un bicchiere freddo.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/l3cd7f1504818306.jpg",
+          "strIngredient1": "Gin",
+          "strIngredient2": "Grenadine",
+          "strIngredient3": "Heavy cream",
+          "strIngredient4": "Milk",
+          "strIngredient5": "Egg White",
+          "strMeasure1": "2 shots ",
+          "strMeasure2": "1/2 shot ",
+          "strMeasure3": "1/2 shot ",
+          "strMeasure4": "1/2 shot",
+          "strMeasure5": "1/2 Fresh",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-07 22:05:06"
         },
         {
-            "idDrink": "17837",
-            "strDrink": "Adam",
-            "strTags": "Alcoholic,Holiday",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "In a shaker half-filled with ice cubes, combine all of the ingredients. Shake well. Strain into a cocktail glass.",
-            "strInstructionsDE": "In einem Shaker, der halb mit Eiswürfeln gefüllt ist, alle Zutaten vermengen. Gut schütteln. In ein Cocktailglas abseihen.",
-            "strInstructionsIT": "In uno shaker riempito a metà con cubetti di ghiaccio, unire tutti gli ingredienti.Agitare bene.Filtrare in un bicchiere da cocktail.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/v0at4i1582478473.jpg",
-            "strIngredient1": "Dark rum",
-            "strIngredient2": "Lemon juice",
-            "strIngredient3": "Grenadine",
-            "strMeasure1": "2 oz ",
-            "strMeasure2": "1 oz ",
-            "strMeasure3": "1 tsp ",
-            "strCreativeCommonsConfirmed": "Yes",
-            "dateModified": "2016-09-02 11:29:29"
+          "idDrink": "17837",
+          "strDrink": "Adam",
+          "strTags": "Alcoholic,Holiday",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "In a shaker half-filled with ice cubes, combine all of the ingredients. Shake well. Strain into a cocktail glass.",
+          "strInstructionsDE": "In einem Shaker, der halb mit Eiswürfeln gefüllt ist, alle Zutaten vermengen. Gut schütteln. In ein Cocktailglas abseihen.",
+          "strInstructionsIT": "In uno shaker riempito a metà con cubetti di ghiaccio, unire tutti gli ingredienti.Agitare bene.Filtrare in un bicchiere da cocktail.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/v0at4i1582478473.jpg",
+          "strIngredient1": "Dark rum",
+          "strIngredient2": "Lemon juice",
+          "strIngredient3": "Grenadine",
+          "strMeasure1": "2 oz ",
+          "strMeasure2": "1 oz ",
+          "strMeasure3": "1 tsp ",
+          "strCreativeCommonsConfirmed": "Yes",
+          "dateModified": "2016-09-02 11:29:29"
         },
         {
-            "idDrink": "13938",
-            "strDrink": "AT&T",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Highball Glass",
-            "strInstructions": "Pour Vodka and Gin over ice, add Tonic and Stir",
-            "strInstructionsDE": "Wodka und Gin über das Eis gießen, Tonic hinzufügen und umrühren.",
-            "strInstructionsIT": "Versare la vodka e il gin sul ghiaccio, aggiungere l'acqua tonica e mescolare",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/rhhwmp1493067619.jpg",
-            "strIngredient1": "Absolut Vodka",
-            "strIngredient2": "Gin",
-            "strIngredient3": "Tonic water",
-            "strMeasure1": "1 oz ",
-            "strMeasure2": "1 oz ",
-            "strMeasure3": "4 oz ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-04-24 22:00:19"
+          "idDrink": "13938",
+          "strDrink": "AT&T",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Highball Glass",
+          "strInstructions": "Pour Vodka and Gin over ice, add Tonic and Stir",
+          "strInstructionsDE": "Wodka und Gin über das Eis gießen, Tonic hinzufügen und umrühren.",
+          "strInstructionsIT": "Versare la vodka e il gin sul ghiaccio, aggiungere l'acqua tonica e mescolare",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/rhhwmp1493067619.jpg",
+          "strIngredient1": "Absolut Vodka",
+          "strIngredient2": "Gin",
+          "strIngredient3": "Tonic water",
+          "strMeasure1": "1 oz ",
+          "strMeasure2": "1 oz ",
+          "strMeasure3": "4 oz ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-04-24 22:00:19"
         },
         {
-            "idDrink": "14610",
-            "strDrink": "ACID",
-            "strCategory": "Shot",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Shot glass",
-            "strInstructions": "Poor in the 151 first followed by the 101 served with a Coke or Dr Pepper chaser.",
-            "strInstructionsDE": "Gering den 151 gefolgt von der 101, die mit einer Cola oder Dr. Pepper Chaser serviert wird.",
-            "strInstructionsIT": "Versa prima il Bacardi 151 seguito dal 101, servito con una Coca-Cola.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/xuxpxt1479209317.jpg",
-            "strIngredient1": "151 proof rum",
-            "strIngredient2": "Wild Turkey",
-            "strMeasure1": "1 oz Bacardi ",
-            "strMeasure2": "1 oz ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2016-11-15 11:28:37"
+          "idDrink": "14610",
+          "strDrink": "ACID",
+          "strCategory": "Shot",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Shot glass",
+          "strInstructions": "Poor in the 151 first followed by the 101 served with a Coke or Dr Pepper chaser.",
+          "strInstructionsDE": "Gering den 151 gefolgt von der 101, die mit einer Cola oder Dr. Pepper Chaser serviert wird.",
+          "strInstructionsIT": "Versa prima il Bacardi 151 seguito dal 101, servito con una Coca-Cola.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/xuxpxt1479209317.jpg",
+          "strIngredient1": "151 proof rum",
+          "strIngredient2": "Wild Turkey",
+          "strMeasure1": "1 oz Bacardi ",
+          "strMeasure2": "1 oz ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2016-11-15 11:28:37"
         },
         {
-            "idDrink": "17833",
-            "strDrink": "A. J.",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "Shake ingredients with ice, strain into a cocktail glass, and serve.",
-            "strInstructionsIT": "Shakerare gli ingredienti con ghiaccio, filtrare in una coppetta da cocktail e servire.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/l74qo91582480316.jpg",
-            "strIngredient1": "Applejack",
-            "strIngredient2": "Grapefruit juice",
-            "strMeasure1": "1 1/2 oz ",
-            "strMeasure2": "1 oz ",
-            "strCreativeCommonsConfirmed": "Yes",
-            "dateModified": "2016-09-02 11:16:58"
+          "idDrink": "17833",
+          "strDrink": "A. J.",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "Shake ingredients with ice, strain into a cocktail glass, and serve.",
+          "strInstructionsIT": "Shakerare gli ingredienti con ghiaccio, filtrare in una coppetta da cocktail e servire.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/l74qo91582480316.jpg",
+          "strIngredient1": "Applejack",
+          "strIngredient2": "Grapefruit juice",
+          "strMeasure1": "1 1/2 oz ",
+          "strMeasure2": "1 oz ",
+          "strCreativeCommonsConfirmed": "Yes",
+          "dateModified": "2016-09-02 11:16:58"
         },
         {
-            "idDrink": "12764",
-            "strDrink": "Karsk",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Highball glass",
-            "strInstructions": "Put a copper coin in a coffe-cup and fill up with coffee until you no longer see the coin, then add alcohol until you see the coin. Norwegian speciality.",
-            "strInstructionsDE": "Legen Sie eine Kupfermünze in eine Kaffeetasse und füllen Sie sie mit Kaffee, bis Sie die Münze nicht mehr sehen, dann fügen Sie Alkohol hinzu, bis Sie die Münze wieder sehen. Norwegische Spezialität.",
-            "strInstructionsIT": "Metti una moneta di rame in una tazza da caffè e riempila di caffè finché non vedi più la moneta, quindi aggiungi l'alcol finché non vedi la moneta. Specialità norvegese.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/808mxk1487602471.jpg",
-            "strIngredient1": "Coffee",
-            "strIngredient2": "Grain alcohol",
-            "strMeasure1": "1 part ",
-            "strMeasure2": "2 parts ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-02-20 14:54:31"
+          "idDrink": "12764",
+          "strDrink": "Karsk",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Highball glass",
+          "strInstructions": "Put a copper coin in a coffe-cup and fill up with coffee until you no longer see the coin, then add alcohol until you see the coin. Norwegian speciality.",
+          "strInstructionsDE": "Legen Sie eine Kupfermünze in eine Kaffeetasse und füllen Sie sie mit Kaffee, bis Sie die Münze nicht mehr sehen, dann fügen Sie Alkohol hinzu, bis Sie die Münze wieder sehen. Norwegische Spezialität.",
+          "strInstructionsIT": "Metti una moneta di rame in una tazza da caffè e riempila di caffè finché non vedi più la moneta, quindi aggiungi l'alcol finché non vedi la moneta. Specialità norvegese.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/808mxk1487602471.jpg",
+          "strIngredient1": "Coffee",
+          "strIngredient2": "Grain alcohol",
+          "strMeasure1": "1 part ",
+          "strMeasure2": "2 parts ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-02-20 14:54:31"
         },
         {
-            "idDrink": "12776",
-            "strDrink": "Melya",
-            "strCategory": "Coffee / Tea",
-            "strAlcoholic": "Non alcoholic",
-            "strGlass": "Coffee mug",
-            "strInstructions": "Brew espresso. In a coffee mug, place 1 teaspoon of unsweetened powdered cocoa, then cover a teaspoon with honey and drizzle it into the cup. Stir while the coffee brews, this is the fun part. The cocoa seems to coat the honey without mixing, so you get a dusty, sticky mass that looks as though it will never mix. Then all at once, presto! It looks like dark chocolate sauce. Pour hot espresso over the honey, stirring to dissolve. Serve with cream.",
-            "strInstructionsDE": "Espresso zubereiten. In eine Kaffeetasse 1 Teelöffel ungesüßten, pulverisierten Kakao geben, dann einen Teelöffel Honig in die Tasse träufeln. Rühren Sie um, während der Kaffee brüht, das ist der lustige Teil. Der Kakao scheint den Honig ohne Mischen zu überziehen, so dass Sie eine dichte, klebrige Masse erhalten, die aussieht, als würde sie sich nie vermischen. Dann auf einmal, presto! Es sieht aus wie dunkle Schokoladensoße. Heißen Espresso über den Honig gießen und unter Rühren auflösen. Mit Sahne servieren.",
-            "strInstructionsIT": "Prepara l'espresso.\r\nIn una tazza da caffè, mettete 1 cucchiaino di cacao amaro in polvere, quindi coprite un cucchiaino con del miele e versatelo nella tazza.\r\nMescola mentre il caffè si prepara, questa è la parte divertente.\r\nIl cacao sembra ricoprire il miele senza mescolarsi, quindi si ottiene una massa polverosa e appiccicosa che sembra non si mescolerà mai.\r\nVersare l'espresso caldo sul miele, mescolando per scioglierlo.\r\nServire con la panna.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/xwtptq1441247579.jpg",
-            "strIngredient1": "Espresso",
-            "strIngredient2": "Honey",
-            "strIngredient3": "Cocoa powder",
-            "strMeasure2": "Unsweetened ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2015-09-03 03:32:59"
+          "idDrink": "12776",
+          "strDrink": "Melya",
+          "strCategory": "Coffee / Tea",
+          "strAlcoholic": "Non alcoholic",
+          "strGlass": "Coffee mug",
+          "strInstructions": "Brew espresso. In a coffee mug, place 1 teaspoon of unsweetened powdered cocoa, then cover a teaspoon with honey and drizzle it into the cup. Stir while the coffee brews, this is the fun part. The cocoa seems to coat the honey without mixing, so you get a dusty, sticky mass that looks as though it will never mix. Then all at once, presto! It looks like dark chocolate sauce. Pour hot espresso over the honey, stirring to dissolve. Serve with cream.",
+          "strInstructionsDE": "Espresso zubereiten. In eine Kaffeetasse 1 Teelöffel ungesüßten, pulverisierten Kakao geben, dann einen Teelöffel Honig in die Tasse träufeln. Rühren Sie um, während der Kaffee brüht, das ist der lustige Teil. Der Kakao scheint den Honig ohne Mischen zu überziehen, so dass Sie eine dichte, klebrige Masse erhalten, die aussieht, als würde sie sich nie vermischen. Dann auf einmal, presto! Es sieht aus wie dunkle Schokoladensoße. Heißen Espresso über den Honig gießen und unter Rühren auflösen. Mit Sahne servieren.",
+          "strInstructionsIT": "Prepara l'espresso.\r\nIn una tazza da caffè, mettete 1 cucchiaino di cacao amaro in polvere, quindi coprite un cucchiaino con del miele e versatelo nella tazza.\r\nMescola mentre il caffè si prepara, questa è la parte divertente.\r\nIl cacao sembra ricoprire il miele senza mescolarsi, quindi si ottiene una massa polverosa e appiccicosa che sembra non si mescolerà mai.\r\nVersare l'espresso caldo sul miele, mescolando per scioglierlo.\r\nServire con la panna.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/xwtptq1441247579.jpg",
+          "strIngredient1": "Espresso",
+          "strIngredient2": "Honey",
+          "strIngredient3": "Cocoa powder",
+          "strMeasure2": "Unsweetened ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2015-09-03 03:32:59"
         },
         {
-            "idDrink": "17839",
-            "strDrink": "Affair",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Highball glass",
-            "strInstructions": "Pour schnapps, orange juice, and cranberry juice over ice in a highball glass. Top with club soda and serve.",
-            "strInstructionsDE": "Gießen Sie Schnaps, Orangensaft und Cranberrysaft über Eis in ein Highball-Glas. Mit Club-Soda überziehen und servieren.",
-            "strInstructionsIT": "Versare la grappa, il succo d'arancia e il succo di mirtillo rosso sul ghiaccio in un bicchiere highball. Completare con la soda club e servire.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/h5za6y1582477994.jpg",
-            "strIngredient1": "Strawberry schnapps",
-            "strIngredient2": "Orange juice",
-            "strIngredient3": "Cranberry juice",
-            "strIngredient4": "Club soda",
-            "strMeasure1": "2 oz ",
-            "strMeasure2": "2 oz ",
-            "strMeasure3": "2 oz ",
-            "strCreativeCommonsConfirmed": "Yes",
-            "dateModified": "2016-09-01 10:05:34"
+          "idDrink": "17839",
+          "strDrink": "Affair",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Highball glass",
+          "strInstructions": "Pour schnapps, orange juice, and cranberry juice over ice in a highball glass. Top with club soda and serve.",
+          "strInstructionsDE": "Gießen Sie Schnaps, Orangensaft und Cranberrysaft über Eis in ein Highball-Glas. Mit Club-Soda überziehen und servieren.",
+          "strInstructionsIT": "Versare la grappa, il succo d'arancia e il succo di mirtillo rosso sul ghiaccio in un bicchiere highball. Completare con la soda club e servire.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/h5za6y1582477994.jpg",
+          "strIngredient1": "Strawberry schnapps",
+          "strIngredient2": "Orange juice",
+          "strIngredient3": "Cranberry juice",
+          "strIngredient4": "Club soda",
+          "strMeasure1": "2 oz ",
+          "strMeasure2": "2 oz ",
+          "strMeasure3": "2 oz ",
+          "strCreativeCommonsConfirmed": "Yes",
+          "dateModified": "2016-09-01 10:05:34"
         },
         {
-            "idDrink": "11149",
-            "strDrink": "Boxcar",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Whiskey sour glass",
-            "strInstructions": "In a shaker half-filled with ice cubes, combine all of the ingredients. Shake well. Strain into a sour glass.",
-            "strInstructionsDE": "In einem Shaker, der halb mit Eiswürfeln gefüllt ist, alle Zutaten vermengen. Gut schütteln. In ein Sour Glas abseihen.",
-            "strInstructionsIT": "In uno shaker riempito a metà con cubetti di ghiaccio, unire tutti gli ingredienti.Filtrare in un bicchiere sour.Agitare bene.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/pwgtpa1504366376.jpg",
-            "strIngredient1": "Gin",
-            "strIngredient2": "Triple sec",
-            "strIngredient3": "Lemon juice",
-            "strIngredient4": "Grenadine",
-            "strIngredient5": "Egg white",
-            "strMeasure1": "1 1/2 oz ",
-            "strMeasure2": "1 oz ",
-            "strMeasure3": "1 tsp ",
-            "strMeasure4": "1/2 tsp ",
-            "strMeasure5": "1 ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-02 16:32:57"
+          "idDrink": "11149",
+          "strDrink": "Boxcar",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Whiskey sour glass",
+          "strInstructions": "In a shaker half-filled with ice cubes, combine all of the ingredients. Shake well. Strain into a sour glass.",
+          "strInstructionsDE": "In einem Shaker, der halb mit Eiswürfeln gefüllt ist, alle Zutaten vermengen. Gut schütteln. In ein Sour Glas abseihen.",
+          "strInstructionsIT": "In uno shaker riempito a metà con cubetti di ghiaccio, unire tutti gli ingredienti.Filtrare in un bicchiere sour.Agitare bene.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/pwgtpa1504366376.jpg",
+          "strIngredient1": "Gin",
+          "strIngredient2": "Triple sec",
+          "strIngredient3": "Lemon juice",
+          "strIngredient4": "Grenadine",
+          "strIngredient5": "Egg white",
+          "strMeasure1": "1 1/2 oz ",
+          "strMeasure2": "1 oz ",
+          "strMeasure3": "1 tsp ",
+          "strMeasure4": "1/2 tsp ",
+          "strMeasure5": "1 ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-02 16:32:57"
         },
         {
-            "idDrink": "11872",
-            "strDrink": "Orgasm",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "Shake all ingredients with ice, strain into a chilled cocktail glass, and serve.",
-            "strInstructionsIT": "Shakerare tutti gli ingredienti con ghiaccio, filtrare in una coppetta da cocktail fredda e servire.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/vr6kle1504886114.jpg",
-            "strIngredient1": "Creme de Cacao",
-            "strIngredient2": "Amaretto",
-            "strIngredient3": "Triple sec",
-            "strIngredient4": "Vodka",
-            "strIngredient5": "Light cream",
-            "strMeasure1": "1/2 oz white ",
-            "strMeasure2": "1/2 oz ",
-            "strMeasure3": "1/2 oz ",
-            "strMeasure4": "1/2 oz ",
-            "strMeasure5": "1 oz ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-08 16:55:14"
+          "idDrink": "11872",
+          "strDrink": "Orgasm",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "Shake all ingredients with ice, strain into a chilled cocktail glass, and serve.",
+          "strInstructionsIT": "Shakerare tutti gli ingredienti con ghiaccio, filtrare in una coppetta da cocktail fredda e servire.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/vr6kle1504886114.jpg",
+          "strIngredient1": "Creme de Cacao",
+          "strIngredient2": "Amaretto",
+          "strIngredient3": "Triple sec",
+          "strIngredient4": "Vodka",
+          "strIngredient5": "Light cream",
+          "strMeasure1": "1/2 oz white ",
+          "strMeasure2": "1/2 oz ",
+          "strMeasure3": "1/2 oz ",
+          "strMeasure4": "1/2 oz ",
+          "strMeasure5": "1 oz ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-08 16:55:14"
         },
         {
-            "idDrink": "15106",
-            "strDrink": "Apello",
-            "strCategory": "Other/Unknown",
-            "strAlcoholic": "Non alcoholic",
-            "strGlass": "Collins Glass",
-            "strInstructions": "Stirr. Grnish with maraschino cherry.",
-            "strInstructionsDE": "Rühren. Mit Maraschinokirsche garnieren.",
-            "strInstructionsIT": "Versare tutti gli ingredienti in un mixing glass. Mescolate.\r\nGuarnire con una ciliegia al maraschino.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/uptxtv1468876415.jpg",
-            "strIngredient1": "Orange juice",
-            "strIngredient2": "Grapefruit juice",
-            "strIngredient3": "Apple juice",
-            "strIngredient4": "Maraschino cherry",
-            "strMeasure1": "4 cl ",
-            "strMeasure2": "3 cl ",
-            "strMeasure3": "1 cl ",
-            "strMeasure4": "1 ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2016-07-18 22:13:35"
+          "idDrink": "15106",
+          "strDrink": "Apello",
+          "strCategory": "Other/Unknown",
+          "strAlcoholic": "Non alcoholic",
+          "strGlass": "Collins Glass",
+          "strInstructions": "Stirr. Grnish with maraschino cherry.",
+          "strInstructionsDE": "Rühren. Mit Maraschinokirsche garnieren.",
+          "strInstructionsIT": "Versare tutti gli ingredienti in un mixing glass. Mescolate.\r\nGuarnire con una ciliegia al maraschino.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/uptxtv1468876415.jpg",
+          "strIngredient1": "Orange juice",
+          "strIngredient2": "Grapefruit juice",
+          "strIngredient3": "Apple juice",
+          "strIngredient4": "Maraschino cherry",
+          "strMeasure1": "4 cl ",
+          "strMeasure2": "3 cl ",
+          "strMeasure3": "1 cl ",
+          "strMeasure4": "1 ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2016-07-18 22:13:35"
         },
         {
-            "idDrink": "15266",
-            "strDrink": "Avalon",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Highball glass",
-            "strInstructions": "Fill a tall glass with ice. Layer the Finlandia Vodka, lemon and apple juices, Pisang Ambon, and top up with lemonade. Stir slightly and garnish with a spiralled cucumber skin and a red cherry. The cucumber provides zest and looks attractive. This drink, created by Timo Haimi, took first prize in the 1991 Finlandia Vodka Long Drink Competition.",
-            "strInstructionsDE": "Füllen Sie ein hohes Glas mit Eis. Legen Sie den Finlandia Wodka, Zitronen- und Apfelsäfte, Pisang Ambon und geben Sie Limonade dazu. Leicht umrühren und mit einer spiralförmigen Gurkenhaut und einer roten Kirsche garnieren. Die Gurke sorgt für Schärfe und sieht attraktiv aus. Dieses von Timo Haimi kreierte Getränk erhielt 1991 den ersten Preis beim Finlandia Wodka Long Drink Wettbewerb.",
-            "strInstructionsIT": "Riempi un bicchiere alto di ghiaccio.\r\nVersare la Vodka, succhi di limone, mela, Pisang Ambon o un liquore alla banana e completare con la limonata.\r\nMescolare leggermente e guarnire con una buccia di cetriolo a spirale e una ciliegia rossa.\r\nIl cetriolo fornisce la scorza e sembra attraente.\r\nQuesta bevanda, creata da Timo Haimi, ha vinto il primo premio nel 1991 Finlandia Vodka Long Drink Competition.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/3k9qic1493068931.jpg",
-            "strIngredient1": "Vodka",
-            "strIngredient2": "Pisang Ambon",
-            "strIngredient3": "Apple juice",
-            "strIngredient4": "Lemon juice",
-            "strIngredient5": "Lemonade",
-            "strMeasure1": "3 parts",
-            "strMeasure2": "1 part ",
-            "strMeasure3": "6 parts ",
-            "strMeasure4": "1 1/2 part ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-04-24 22:22:11"
+          "idDrink": "15266",
+          "strDrink": "Avalon",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Highball glass",
+          "strInstructions": "Fill a tall glass with ice. Layer the Finlandia Vodka, lemon and apple juices, Pisang Ambon, and top up with lemonade. Stir slightly and garnish with a spiralled cucumber skin and a red cherry. The cucumber provides zest and looks attractive. This drink, created by Timo Haimi, took first prize in the 1991 Finlandia Vodka Long Drink Competition.",
+          "strInstructionsDE": "Füllen Sie ein hohes Glas mit Eis. Legen Sie den Finlandia Wodka, Zitronen- und Apfelsäfte, Pisang Ambon und geben Sie Limonade dazu. Leicht umrühren und mit einer spiralförmigen Gurkenhaut und einer roten Kirsche garnieren. Die Gurke sorgt für Schärfe und sieht attraktiv aus. Dieses von Timo Haimi kreierte Getränk erhielt 1991 den ersten Preis beim Finlandia Wodka Long Drink Wettbewerb.",
+          "strInstructionsIT": "Riempi un bicchiere alto di ghiaccio.\r\nVersare la Vodka, succhi di limone, mela, Pisang Ambon o un liquore alla banana e completare con la limonata.\r\nMescolare leggermente e guarnire con una buccia di cetriolo a spirale e una ciliegia rossa.\r\nIl cetriolo fornisce la scorza e sembra attraente.\r\nQuesta bevanda, creata da Timo Haimi, ha vinto il primo premio nel 1991 Finlandia Vodka Long Drink Competition.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/3k9qic1493068931.jpg",
+          "strIngredient1": "Vodka",
+          "strIngredient2": "Pisang Ambon",
+          "strIngredient3": "Apple juice",
+          "strIngredient4": "Lemon juice",
+          "strIngredient5": "Lemonade",
+          "strMeasure1": "3 parts",
+          "strMeasure2": "1 part ",
+          "strMeasure3": "6 parts ",
+          "strMeasure4": "1 1/2 part ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-04-24 22:22:11"
         },
         {
-            "idDrink": "17185",
-            "strDrink": "Casino",
-            "strTags": "IBA",
-            "strVideo": "https://www.youtube.com/watch?v=C6v8NsaDNj8",
-            "strCategory": "Ordinary Drink",
-            "strIBA": "Unforgettables",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "Pour all ingredients into shaker with ice cubes. Shake well. Strain into chilled cocktail glass. Garnish with a lemon twist and a maraschino cherry. Serve without a straw.",
-            "strInstructionsDE": "Alle Zutaten in den Shaker mit Eiswürfeln geben. Gut schütteln. In ein gekühltes Cocktailglas abseihen. Mit einer Zitronenscheibe und einer Maraschinokirsche garnieren. Ohne Strohhalm servieren.",
-            "strInstructionsIT": "Versare tutti gli ingredienti in uno shaker con cubetti di ghiaccio.\r\nAgitare bene.\r\nFiltrare in una coppetta da cocktail ghiacciata.\r\nGuarnire con una scorza di limone e una ciliegia al maraschino.\r\nServire senza cannuccia.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/1mvjxg1504348579.jpg",
-            "strIngredient1": "Gin",
-            "strIngredient2": "Maraschino liqueur",
-            "strIngredient3": "Lemon juice",
-            "strIngredient4": "Orange bitters",
-            "strIngredient5": "Cherry",
-            "strMeasure1": "2 oz ",
-            "strMeasure2": "1/4 tsp ",
-            "strMeasure3": "1/4 tsp ",
-            "strMeasure4": "2 dashes ",
-            "strMeasure5": "1 ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-02 11:36:19"
+          "idDrink": "17185",
+          "strDrink": "Casino",
+          "strTags": "IBA",
+          "strVideo": "https://www.youtube.com/watch?v=C6v8NsaDNj8",
+          "strCategory": "Ordinary Drink",
+          "strIBA": "Unforgettables",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "Pour all ingredients into shaker with ice cubes. Shake well. Strain into chilled cocktail glass. Garnish with a lemon twist and a maraschino cherry. Serve without a straw.",
+          "strInstructionsDE": "Alle Zutaten in den Shaker mit Eiswürfeln geben. Gut schütteln. In ein gekühltes Cocktailglas abseihen. Mit einer Zitronenscheibe und einer Maraschinokirsche garnieren. Ohne Strohhalm servieren.",
+          "strInstructionsIT": "Versare tutti gli ingredienti in uno shaker con cubetti di ghiaccio.\r\nAgitare bene.\r\nFiltrare in una coppetta da cocktail ghiacciata.\r\nGuarnire con una scorza di limone e una ciliegia al maraschino.\r\nServire senza cannuccia.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/1mvjxg1504348579.jpg",
+          "strIngredient1": "Gin",
+          "strIngredient2": "Maraschino liqueur",
+          "strIngredient3": "Lemon juice",
+          "strIngredient4": "Orange bitters",
+          "strIngredient5": "Cherry",
+          "strMeasure1": "2 oz ",
+          "strMeasure2": "1/4 tsp ",
+          "strMeasure3": "1/4 tsp ",
+          "strMeasure4": "2 dashes ",
+          "strMeasure5": "1 ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-02 11:36:19"
         },
         {
-            "idDrink": "16031",
-            "strDrink": "Radler",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Highball glass",
-            "strInstructions": "Pour beer into large mug, slowly add the 7-up (or Sprite).",
-            "strInstructionsDE": "Bier in einen großen Becher gießen, langsam den 7-up (oder Sprite) hinzufügen.",
-            "strInstructionsIT": "Versare la birra in una tazza grande, aggiungere lentamente il 7up (o Sprite).",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/xz8igv1504888995.jpg",
-            "strIngredient1": "Beer",
-            "strIngredient2": "7-Up",
-            "strMeasure1": "12 oz ",
-            "strMeasure2": "12 oz ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-08 17:43:15"
+          "idDrink": "16031",
+          "strDrink": "Radler",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Highball glass",
+          "strInstructions": "Pour beer into large mug, slowly add the 7-up (or Sprite).",
+          "strInstructionsDE": "Bier in einen großen Becher gießen, langsam den 7-up (oder Sprite) hinzufügen.",
+          "strInstructionsIT": "Versare la birra in una tazza grande, aggiungere lentamente il 7up (o Sprite).",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/xz8igv1504888995.jpg",
+          "strIngredient1": "Beer",
+          "strIngredient2": "7-Up",
+          "strMeasure1": "12 oz ",
+          "strMeasure2": "12 oz ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-08 17:43:15"
         },
         {
-            "idDrink": "17205",
-            "strDrink": "Mimosa",
-            "strTags": "IBA,ContemporaryClassic",
-            "strCategory": "Ordinary Drink",
-            "strIBA": "Contemporary Classics",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Champagne flute",
-            "strInstructions": "Ensure both ingredients are well chilled, then mix into the glass. Serve cold.",
-            "strInstructionsDE": "Achten Sie darauf, dass beide Zutaten gut gekühlt sind, und mischen Sie sie dann in das Glas. Kalt servieren.",
-            "strInstructionsIT": "Assicurati che entrambi gli ingredienti siano ben raffreddati, quindi mescolali nel bicchiere.\r\nServire freddo.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/juhcuu1504370685.jpg",
-            "strIngredient1": "Champagne",
-            "strIngredient2": "Orange juice",
-            "strMeasure1": "Chilled ",
-            "strMeasure2": "2 oz ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-02 17:44:45"
+          "idDrink": "17205",
+          "strDrink": "Mimosa",
+          "strTags": "IBA,ContemporaryClassic",
+          "strCategory": "Ordinary Drink",
+          "strIBA": "Contemporary Classics",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Champagne flute",
+          "strInstructions": "Ensure both ingredients are well chilled, then mix into the glass. Serve cold.",
+          "strInstructionsDE": "Achten Sie darauf, dass beide Zutaten gut gekühlt sind, und mischen Sie sie dann in das Glas. Kalt servieren.",
+          "strInstructionsIT": "Assicurati che entrambi gli ingredienti siano ben raffreddati, quindi mescolali nel bicchiere.\r\nServire freddo.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/juhcuu1504370685.jpg",
+          "strIngredient1": "Champagne",
+          "strIngredient2": "Orange juice",
+          "strMeasure1": "Chilled ",
+          "strMeasure2": "2 oz ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-02 17:44:45"
         },
         {
-            "idDrink": "17253",
-            "strDrink": "Paloma",
-            "strCategory": "Cocktail",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Collins glass",
-            "strInstructions": "Stir together and serve over ice.",
-            "strInstructionsDE": "Zusammenrühren und über Eis servieren.",
-            "strInstructionsIT": "Mescolate e servite con ghiaccio.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/samm5j1513706393.jpg",
-            "strIngredient1": "Grape Soda",
-            "strIngredient2": "Tequila",
-            "strMeasure1": "3 oz",
-            "strMeasure2": "1 1/2 oz",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-12-19 17:59:53"
+          "idDrink": "17253",
+          "strDrink": "Paloma",
+          "strCategory": "Cocktail",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Collins glass",
+          "strInstructions": "Stir together and serve over ice.",
+          "strInstructionsDE": "Zusammenrühren und über Eis servieren.",
+          "strInstructionsIT": "Mescolate e servite con ghiaccio.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/samm5j1513706393.jpg",
+          "strIngredient1": "Grape Soda",
+          "strIngredient2": "Tequila",
+          "strMeasure1": "3 oz",
+          "strMeasure2": "1 1/2 oz",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-12-19 17:59:53"
         },
         {
-            "idDrink": "17835",
-            "strDrink": "Abilene",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Highball glass",
-            "strInstructions": "Pour all of the ingredients into a highball glass almost filled with ice cubes. Stir well.",
-            "strInstructionsIT": "Versare tutti gli ingredienti in un bicchiere highball riempito di cubetti di ghiaccio. Mescolare bene.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/smb2oe1582479072.jpg",
-            "strIngredient1": "Dark rum",
-            "strIngredient2": "Peach nectar",
-            "strIngredient3": "Orange juice",
-            "strMeasure1": "1 1/2 oz ",
-            "strMeasure2": "2 oz ",
-            "strMeasure3": "3 oz ",
-            "strCreativeCommonsConfirmed": "Yes",
-            "dateModified": "2016-04-29 09:44:25"
+          "idDrink": "17835",
+          "strDrink": "Abilene",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Highball glass",
+          "strInstructions": "Pour all of the ingredients into a highball glass almost filled with ice cubes. Stir well.",
+          "strInstructionsIT": "Versare tutti gli ingredienti in un bicchiere highball riempito di cubetti di ghiaccio. Mescolare bene.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/smb2oe1582479072.jpg",
+          "strIngredient1": "Dark rum",
+          "strIngredient2": "Peach nectar",
+          "strIngredient3": "Orange juice",
+          "strMeasure1": "1 1/2 oz ",
+          "strMeasure2": "2 oz ",
+          "strMeasure3": "3 oz ",
+          "strCreativeCommonsConfirmed": "Yes",
+          "dateModified": "2016-04-29 09:44:25"
         },
         {
-            "idDrink": "11023",
-            "strDrink": "Almeria",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "In a shaker half-filled with ice cubes, combine all of the ingredients. Shake well. Strain into a cocktail glass.",
-            "strInstructionsDE": "In einem Shaker, der halb mit Eiswürfeln gefüllt ist, alle Zutaten vermengen. Gut schütteln. In ein Cocktailglas abseihen.",
-            "strInstructionsIT": "In uno shaker riempito a metà con cubetti di ghiaccio, unire tutti gli ingredienti.Agitare bene.Filtrare in un bicchiere da cocktail.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/rwsyyu1483388181.jpg",
-            "strIngredient1": "Dark rum",
-            "strIngredient2": "Kahlua",
-            "strIngredient3": "Egg white",
-            "strMeasure1": "2 oz ",
-            "strMeasure2": "1 oz ",
-            "strMeasure3": "1 ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-01-02 20:16:21"
+          "idDrink": "11023",
+          "strDrink": "Almeria",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "In a shaker half-filled with ice cubes, combine all of the ingredients. Shake well. Strain into a cocktail glass.",
+          "strInstructionsDE": "In einem Shaker, der halb mit Eiswürfeln gefüllt ist, alle Zutaten vermengen. Gut schütteln. In ein Cocktailglas abseihen.",
+          "strInstructionsIT": "In uno shaker riempito a metà con cubetti di ghiaccio, unire tutti gli ingredienti.Agitare bene.Filtrare in un bicchiere da cocktail.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/rwsyyu1483388181.jpg",
+          "strIngredient1": "Dark rum",
+          "strIngredient2": "Kahlua",
+          "strIngredient3": "Egg white",
+          "strMeasure1": "2 oz ",
+          "strMeasure2": "1 oz ",
+          "strMeasure3": "1 ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-01-02 20:16:21"
         },
         {
-            "idDrink": "11690",
-            "strDrink": "Mai Tai",
-            "strTags": "IBA,ContemporaryClassic",
-            "strCategory": "Ordinary Drink",
-            "strIBA": "Contemporary Classics",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Collins glass",
-            "strInstructions": "Shake all ingredients with ice. Strain into glass. Garnish and serve with straw.",
-            "strInstructionsDE": "Alle Zutaten mit Eis schütteln. In ein Glas abseihen. Garnieren und mit Trinkhalm servieren.",
-            "strInstructionsIT": "Shakerare tutti gli ingredienti con ghiaccio.Filtrare nel bicchiere. ",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/twyrrp1439907470.jpg",
-            "strIngredient1": "Light rum",
-            "strIngredient2": "Orgeat syrup",
-            "strIngredient3": "Triple sec",
-            "strIngredient4": "Sweet and sour",
-            "strIngredient5": "Cherry",
-            "strMeasure1": "1 oz ",
-            "strMeasure2": "1/2 oz ",
-            "strMeasure3": "1/2 oz ",
-            "strMeasure4": "1 1/2 oz ",
-            "strMeasure5": "1 ",
-            "strImageSource": "https://commons.wikimedia.org/wiki/File:Mai_Tai_(16304400706).jpg",
-            "strImageAttribution": "Johnny Silvercloud https://www.flickr.com/people/116193477@N08",
-            "strCreativeCommonsConfirmed": "Yes",
-            "dateModified": "2015-08-18 15:17:50"
+          "idDrink": "11690",
+          "strDrink": "Mai Tai",
+          "strTags": "IBA,ContemporaryClassic",
+          "strCategory": "Ordinary Drink",
+          "strIBA": "Contemporary Classics",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Collins glass",
+          "strInstructions": "Shake all ingredients with ice. Strain into glass. Garnish and serve with straw.",
+          "strInstructionsDE": "Alle Zutaten mit Eis schütteln. In ein Glas abseihen. Garnieren und mit Trinkhalm servieren.",
+          "strInstructionsIT": "Shakerare tutti gli ingredienti con ghiaccio.Filtrare nel bicchiere. ",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/twyrrp1439907470.jpg",
+          "strIngredient1": "Light rum",
+          "strIngredient2": "Orgeat syrup",
+          "strIngredient3": "Triple sec",
+          "strIngredient4": "Sweet and sour",
+          "strIngredient5": "Cherry",
+          "strMeasure1": "1 oz ",
+          "strMeasure2": "1/2 oz ",
+          "strMeasure3": "1/2 oz ",
+          "strMeasure4": "1 1/2 oz ",
+          "strMeasure5": "1 ",
+          "strImageSource": "https://commons.wikimedia.org/wiki/File:Mai_Tai_(16304400706).jpg",
+          "strImageAttribution": "Johnny Silvercloud https://www.flickr.com/people/116193477@N08",
+          "strCreativeCommonsConfirmed": "Yes",
+          "dateModified": "2015-08-18 15:17:50"
         },
         {
-            "idDrink": "11728",
-            "strDrink": "Martini",
-            "strVideo": "https://www.youtube.com/watch?v=ApMR3IWYZHI",
-            "strCategory": "Cocktail",
-            "strIBA": "Unforgettables",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "Straight: Pour all ingredients into mixing glass with ice cubes. Stir well. Strain in chilled martini cocktail glass. Squeeze oil from lemon peel onto the drink, or garnish with olive.",
-            "strInstructionsDE": "Direkt: Alle Zutaten in ein Mischglas mit Eiswürfeln geben. Gut umrühren. In einem gekühlten Martini-Cocktailglas abseihen. Den Saft aus der Zitronenschale auf das Getränk drücken oder mit Olive garnieren.",
-            "strInstructionsIT": "Versare tutti gli ingredienti nel mixing glass con cubetti di ghiaccio. Mescolare bene. Spremi la scorza di limone sulla bevanda o guarnisci con l'oliva.Filtrare in una coppetta da cocktail Martini ghiacciata. ",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/71t8581504353095.jpg",
-            "strIngredient1": "Gin",
-            "strIngredient2": "Dry Vermouth",
-            "strIngredient3": "Olive",
-            "strMeasure1": "1 2/3 oz ",
-            "strMeasure2": "1/3 oz ",
-            "strMeasure3": "1 ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-09-02 12:51:35"
+          "idDrink": "11728",
+          "strDrink": "Martini",
+          "strVideo": "https://www.youtube.com/watch?v=ApMR3IWYZHI",
+          "strCategory": "Cocktail",
+          "strIBA": "Unforgettables",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "Straight: Pour all ingredients into mixing glass with ice cubes. Stir well. Strain in chilled martini cocktail glass. Squeeze oil from lemon peel onto the drink, or garnish with olive.",
+          "strInstructionsDE": "Direkt: Alle Zutaten in ein Mischglas mit Eiswürfeln geben. Gut umrühren. In einem gekühlten Martini-Cocktailglas abseihen. Den Saft aus der Zitronenschale auf das Getränk drücken oder mit Olive garnieren.",
+          "strInstructionsIT": "Versare tutti gli ingredienti nel mixing glass con cubetti di ghiaccio. Mescolare bene. Spremi la scorza di limone sulla bevanda o guarnisci con l'oliva.Filtrare in una coppetta da cocktail Martini ghiacciata. ",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/71t8581504353095.jpg",
+          "strIngredient1": "Gin",
+          "strIngredient2": "Dry Vermouth",
+          "strIngredient3": "Olive",
+          "strMeasure1": "1 2/3 oz ",
+          "strMeasure2": "1/3 oz ",
+          "strMeasure3": "1 ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-09-02 12:51:35"
         },
         {
-            "idDrink": "12127",
-            "strDrink": "Sazerac",
-            "strTags": "IBA,Classic",
-            "strVideo": "https://www.youtube.com/watch?v=RuZzOcjJUYw",
-            "strCategory": "Ordinary Drink",
-            "strIBA": "Unforgettables",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Old-fashioned glass",
-            "strInstructions": "Rinse a chilled old-fashioned glass with the absinthe, add crushed ice, and set it aside. Stir the remaining ingredients over ice and set it aside. Discard the ice and any excess absinthe from the prepared glass, and strain the drink into the glass. Add the lemon peel for garnish.",
-            "strInstructionsDE": "Ein gekühltes old-fashioned Glas mit dem Absinth ausspülen, zerstoßenes Eis hinzufügen und beiseite stellen. Die restlichen Zutaten über das Eis geben und beiseite stellen. Entsorgen Sie das Eis und den überschüssigen Absinth aus dem vorbereiteten Glas und seihen Sie das Getränk in das Glas. Die Zitronenschale als Garnierung hinzufügen.",
-            "strInstructionsIT": "Sciacquare un bicchiere vecchio stile freddo con l'assenzio, aggiungere il ghiaccio tritato e metterlo da parte. Mescola gli ingredienti rimanenti sul ghiaccio e mettilo da parte. Togli il ghiaccio e l'assenzio in eccesso dal bicchiere precedentemente preparato e filtra la bevanda nel bicchiere .Aggiungere la scorza di limone per guarnire.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/vvpxwy1439907208.jpg",
-            "strIngredient1": "Ricard",
-            "strIngredient2": "Sugar",
-            "strIngredient3": "Peychaud bitters",
-            "strIngredient4": "Water",
-            "strIngredient5": "Bourbon",
-            "strIngredient6": "Lemon peel",
-            "strMeasure1": "1 tsp ",
-            "strMeasure2": "1/2 tsp superfine ",
-            "strMeasure3": "2 dashes ",
-            "strMeasure4": "1 tsp ",
-            "strMeasure5": "2 oz ",
-            "strMeasure6": "1 twist of ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2015-08-18 15:13:28"
+          "idDrink": "12127",
+          "strDrink": "Sazerac",
+          "strTags": "IBA,Classic",
+          "strVideo": "https://www.youtube.com/watch?v=RuZzOcjJUYw",
+          "strCategory": "Ordinary Drink",
+          "strIBA": "Unforgettables",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Old-fashioned glass",
+          "strInstructions": "Rinse a chilled old-fashioned glass with the absinthe, add crushed ice, and set it aside. Stir the remaining ingredients over ice and set it aside. Discard the ice and any excess absinthe from the prepared glass, and strain the drink into the glass. Add the lemon peel for garnish.",
+          "strInstructionsDE": "Ein gekühltes old-fashioned Glas mit dem Absinth ausspülen, zerstoßenes Eis hinzufügen und beiseite stellen. Die restlichen Zutaten über das Eis geben und beiseite stellen. Entsorgen Sie das Eis und den überschüssigen Absinth aus dem vorbereiteten Glas und seihen Sie das Getränk in das Glas. Die Zitronenschale als Garnierung hinzufügen.",
+          "strInstructionsIT": "Sciacquare un bicchiere vecchio stile freddo con l'assenzio, aggiungere il ghiaccio tritato e metterlo da parte. Mescola gli ingredienti rimanenti sul ghiaccio e mettilo da parte. Togli il ghiaccio e l'assenzio in eccesso dal bicchiere precedentemente preparato e filtra la bevanda nel bicchiere .Aggiungere la scorza di limone per guarnire.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/vvpxwy1439907208.jpg",
+          "strIngredient1": "Ricard",
+          "strIngredient2": "Sugar",
+          "strIngredient3": "Peychaud bitters",
+          "strIngredient4": "Water",
+          "strIngredient5": "Bourbon",
+          "strIngredient6": "Lemon peel",
+          "strMeasure1": "1 tsp ",
+          "strMeasure2": "1/2 tsp superfine ",
+          "strMeasure3": "2 dashes ",
+          "strMeasure4": "1 tsp ",
+          "strMeasure5": "2 oz ",
+          "strMeasure6": "1 twist of ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2015-08-18 15:13:28"
         },
         {
-            "idDrink": "12196",
-            "strDrink": "Sidecar",
-            "strTags": "IBA,Classic",
-            "strVideo": "https://www.youtube.com/watch?v=uvKKILoO-Nk",
-            "strCategory": "Ordinary Drink",
-            "strIBA": "Unforgettables",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Cocktail glass",
-            "strInstructions": "Pour all ingredients into cocktail shaker filled with ice. Shake well and strain into cocktail glass.",
-            "strInstructionsDE": "Alle Zutaten in einen mit Eis gefüllten Cocktailshaker geben. Gut schütteln und in ein Cocktailglas abseihen.",
-            "strInstructionsIT": "Versare tutti gli ingredienti in uno shaker pieno di ghiaccio.Shakerare bene e filtrare nella coppetta da cocktail.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/x72sik1606854964.jpg",
-            "strIngredient1": "Cognac",
-            "strIngredient2": "Cointreau",
-            "strIngredient3": "Lemon juice",
-            "strMeasure1": "2 oz ",
-            "strMeasure2": "1/2 oz ",
-            "strMeasure3": "1 oz ",
-            "strImageSource": "https://www.flickr.com/photos/mstyne/29716525626/in/dateposted/",
-            "strImageAttribution": "Michael Styne https://www.flickr.com/photos/mstyne/",
-            "strCreativeCommonsConfirmed": "Yes",
-            "dateModified": "2015-08-18 15:07:32"
+          "idDrink": "12196",
+          "strDrink": "Sidecar",
+          "strTags": "IBA,Classic",
+          "strVideo": "https://www.youtube.com/watch?v=uvKKILoO-Nk",
+          "strCategory": "Ordinary Drink",
+          "strIBA": "Unforgettables",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Cocktail glass",
+          "strInstructions": "Pour all ingredients into cocktail shaker filled with ice. Shake well and strain into cocktail glass.",
+          "strInstructionsDE": "Alle Zutaten in einen mit Eis gefüllten Cocktailshaker geben. Gut schütteln und in ein Cocktailglas abseihen.",
+          "strInstructionsIT": "Versare tutti gli ingredienti in uno shaker pieno di ghiaccio.Shakerare bene e filtrare nella coppetta da cocktail.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/x72sik1606854964.jpg",
+          "strIngredient1": "Cognac",
+          "strIngredient2": "Cointreau",
+          "strIngredient3": "Lemon juice",
+          "strMeasure1": "2 oz ",
+          "strMeasure2": "1/2 oz ",
+          "strMeasure3": "1 oz ",
+          "strImageSource": "https://www.flickr.com/photos/mstyne/29716525626/in/dateposted/",
+          "strImageAttribution": "Michael Styne https://www.flickr.com/photos/mstyne/",
+          "strCreativeCommonsConfirmed": "Yes",
+          "dateModified": "2015-08-18 15:07:32"
         },
         {
-            "idDrink": "12446",
-            "strDrink": "Veteran",
-            "strCategory": "Ordinary Drink",
-            "strAlcoholic": "Alcoholic",
-            "strGlass": "Old-fashioned glass",
-            "strInstructions": "Pour the rum and cherry brandy into an old-fashioned glass almost filled with ice cubes. Stir well.",
-            "strInstructionsDE": "Den Rum und Kirschlikör in ein old-fashioned Glas gießen, das fast mit Eiswürfeln gefüllt ist. Gut umrühren.",
-            "strInstructionsIT": "Versare il rum e il brandy di ciliegie in un bicchiere vecchio stile riempito per 2/3 di cubetti di ghiaccio. Mescolare bene.",
-            "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/iwml9t1492976255.jpg",
-            "strIngredient1": "Dark rum",
-            "strIngredient2": "Cherry brandy",
-            "strMeasure1": "2 oz ",
-            "strMeasure2": "1/2 oz ",
-            "strCreativeCommonsConfirmed": "No",
-            "dateModified": "2017-04-23 20:37:36"
+          "idDrink": "12446",
+          "strDrink": "Veteran",
+          "strCategory": "Ordinary Drink",
+          "strAlcoholic": "Alcoholic",
+          "strGlass": "Old-fashioned glass",
+          "strInstructions": "Pour the rum and cherry brandy into an old-fashioned glass almost filled with ice cubes. Stir well.",
+          "strInstructionsDE": "Den Rum und Kirschlikör in ein old-fashioned Glas gießen, das fast mit Eiswürfeln gefüllt ist. Gut umrühren.",
+          "strInstructionsIT": "Versare il rum e il brandy di ciliegie in un bicchiere vecchio stile riempito per 2/3 di cubetti di ghiaccio. Mescolare bene.",
+          "strDrinkThumb": "https://www.thecocktaildb.com/images/media/drink/iwml9t1492976255.jpg",
+          "strIngredient1": "Dark rum",
+          "strIngredient2": "Cherry brandy",
+          "strMeasure1": "2 oz ",
+          "strMeasure2": "1/2 oz ",
+          "strCreativeCommonsConfirmed": "No",
+          "dateModified": "2017-04-23 20:37:36"
         }
-    ]
+      ]
 }
